clean up constructor
