+++ ronswanson/simulation_builder.py
@@ -5,10 +5,11 @@ from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
+import h5py
 import numpy as np
 import yaml
-from astromodels.functions.template_model import h5py
 
+from .grids import ParameterGrid
 from .script_generator import (
     PythonGenerator,
     SLURMGatherGenerator,
@@ -20,359 +21,29 @@ log = setup_logger(__name__)
 
 
 @dataclass(frozen=True)
-class EnergyGrid:
+class SLURMTime:
+    hrs: int = 0
+    min: int = 10
+    sec: int = 0
 
-    vmin: Optional[float] = None
-    vmax: Optional[float] = None
-    scale: Optional[str] = None
-    n_points: Optional[int] = None
-    values: Optional[np.ndarray] = None
-    custom: bool = False
 
-    def __post_init__(self):
+@dataclass
+class JobConfig:
+    n_cores_per_node: int
+    time: SLURMTime
 
-        if not self.custom:
 
-            # we will build a grid
+@dataclass
+class GatherConfig(JobConfig):
+    n_gather_per_core: int
 
-            if (self.vmin is None) or (self.vmax is None):
 
-                log.error("non-custom grids must include vmin and vmax")
-
-                raise AssertionError
-
-            if self.scale is None:
-
-                log.error(
-                    "non-custom grids must include scale 'log' or 'linear'"
-                )
-
-                raise AssertionError
-
-            else:
-
-                if self.scale not in ['log', 'linear']:
-
-                    log.error(
-                        "non-custom grids must include scale 'log' or 'linear'"
-                    )
-
-                    raise AssertionError
-
-            if self.n_points is None:
-
-                log.error("non-custom grids must include n_points")
-
-                raise AssertionError
-
-        else:
-
-            if self.values is None:
-
-                log.error("custom grids must include values")
-
-                raise AssertionError
-
-    @property
-    def grid(self) -> np.ndarray:
-
-        if self.custom:
-
-            return self.values
-
-        else:
-
-            if self.scale.lower() == 'log':
-
-                return np.geomspace(self.vmin, self.vmax, self.n_points)
-
-            else:
-
-                return np.linspace(self.vmin, self.vmax, self.n_points)
-
-    @classmethod
-    def from_dict(cls, d: Dict[str, Any]) -> "Parameter":
-
-        inputs = {}
-        inputs["custom"] = d["custom"]
-
-        if d["custom"]:
-
-            inputs["values"] = np.array(d["values"])
-
-        else:
-
-            inputs["vmin"] = d["vmin"]
-            inputs["vmax"] = d["vmax"]
-            inputs["scale"] = d["scale"]
-            inputs["n_points"] = d["n_points"]
-
-        return cls(**inputs)
-
-    def to_dict(self) -> Dict[str, Any]:
-
-        out = dict(custom=self.custom)
-
-        if self.custom:
-
-            out["values"] = self.values.tolist()
-
-        else:
-
-            out["vmin"] = self.vmin
-            out["vmax"] = self.vmax
-            out["scale"] = self.scale
-            out["n_points"] = self.n_points
-
-        return out
-
-
-@dataclass(frozen=True)
-class Parameter:
-
-    name: str
-    vmin: Optional[float] = None
-    vmax: Optional[float] = None
-    scale: Optional[str] = None
-    n_points: Optional[int] = None
-    values: Optional[np.ndarray] = None
-    custom: bool = False
-    #   grid: np.ndarray = field(init=False)
-
-    def __post_init__(self):
-
-        if not self.custom:
-
-            # we will build a grid
-
-            if (self.vmin is None) or (self.vmax is None):
-
-                log.error("non-custom grids must include vmin and vmax")
-
-                raise AssertionError
-
-            if self.scale is None:
-
-                log.error(
-                    "non-custom grids must include scale 'log' or 'linear'"
-                )
-
-                raise AssertionError
-
-            else:
-
-                if self.scale not in ['log', 'linear']:
-
-                    log.error(
-                        "non-custom grids must include scale 'log' or 'linear'"
-                    )
-
-                    raise AssertionError
-
-            if self.n_points is None:
-
-                log.error("non-custom grids must include n_points")
-
-                raise AssertionError
-
-        else:
-
-            if self.values is None:
-
-                log.error("custom grids must include values")
-
-                raise AssertionError
-
-    @property
-    def grid(self) -> np.ndarray:
-
-        if self.custom:
-
-            return self.values
-
-        else:
-
-            if self.scale.lower() == 'log':
-
-                return np.geomspace(self.vmin, self.vmax, self.n_points)
-
-            else:
-
-                return np.linspace(self.vmin, self.vmax, self.n_points)
-
-    @classmethod
-    def from_dict(cls, name: str, d: Dict[str, Any]) -> "Parameter":
-
-        inputs = {}
-        inputs["custom"] = d["custom"]
-
-        if d["custom"]:
-
-            inputs["values"] = np.array(d["values"])
-
-        else:
-
-            inputs["vmin"] = d["vmin"]
-            inputs["vmax"] = d["vmax"]
-            inputs["scale"] = d["scale"]
-            inputs["n_points"] = d["n_points"]
-
-        return cls(name, **inputs)
-
-    def to_dict(self) -> Dict[str, Any]:
-
-        out = dict(custom=self.custom)
-
-        if self.custom:
-
-            out["values"] = self.values.tolist()
-
-        else:
-
-            out["vmin"] = self.vmin
-            out["vmax"] = self.vmax
-            out["scale"] = self.scale
-            out["n_points"] = self.n_points
-
-        return out
-
-
-@dataclass(frozen=True)
-class ParameterGrid:
-
-    parameter_list: List[Parameter]
-    energy_grid: List[EnergyGrid]
-
-    @property
-    def n_points(self) -> int:
-
-        n = 1
-        for param in self.parameter_list:
-
-            n *= len(param.grid)
-
-        return n
-
-    @property
-    def n_parameters(self) -> int:
-
-        return len(self.parameter_list)
-
-    @classmethod
-    def from_dict(cls, d: Dict[str, Dict[str, Any]]) -> "ParameterGrid":
-
-        # make sure to sort so that we always have the same parameter
-        # ordering
-
-        is_multi_output: bool = False
-        n_energy_grids = 1
-
-        for k in d.keys():
-
-            if "energy_grid" in k:
-
-                if len(k.split("_")) == 3:
-
-                    if is_multi_output:
-
-                        # we have already detected one
-
-                        n_energy_grids += 1
-
-                    is_multi_output = True
-
-        if not is_multi_output:
-
-            energy_grid = [EnergyGrid.from_dict(d.pop("energy_grid"))]
-
-        else:
-
-            energy_grid = [
-                EnergyGrid.from_dict(d.pop(f"energy_grid_{i}"))
-                for i in range(n_energy_grids)
-            ]
-
-        pars = list(d.keys())
-
-        pars.sort()
-
-        par_list = [
-            Parameter.from_dict(par_name, d[par_name]) for par_name in pars
-        ]
-
-        return cls(par_list, energy_grid)
-
-    @classmethod
-    def from_yaml(cls, file_name: str) -> "ParameterGrid":
-
-        with open(file_name, 'r') as f:
-
-            inputs = yaml.load(stream=f, Loader=yaml.SafeLoader)
-
-        return cls.from_dict(inputs)
-
-    @property
-    def parameter_names(self) -> List[str]:
-
-        return [p.name for p in self.parameter_list]
-
-    def to_dict(self) -> Dict[str, Dict[str, Any]]:
-
-        out = {}
-
-        for p in self.parameter_list:
-
-            out[p.name] = p.to_dict()
-
-        if len(self.energy_grid) == 1:
-
-            out['energy_grid'] = self.energy_grid[0].to_dict()
-
-        else:
-
-            for i, eg in enumerate(self.energy_grid):
-
-                out[f"energy_grid_{i}"] = eg.to_dict()
-
-        return out
-
-    def write(self, file_name: str) -> None:
-
-        with open(file_name, "w") as f:
-
-            yaml.dump(
-                stream=f,
-                data=self.to_dict(),
-                default_flow_style=False,
-                Dumper=yaml.SafeDumper,
-            )
-
-    def at_index(self, i: int) -> Dict[str, float]:
-        """
-        return the ith set of parameters
-
-        :param i:
-        :type i: int
-        :returns:
-
-        """
-        idx = 0
-
-        for result in itertools.product(*[p.grid for p in self.parameter_list]):
-
-            if i == idx:
-
-                d = OrderedDict()
-
-                for k, v in zip(self.parameter_names, result):
-
-                    d[k] = v
-
-                return d
-
-            else:
-
-                idx += 1
+@dataclass
+class SimulationConfig(JobConfig):
+    n_mp_jobs: int
+    run_per_node: int = 1
+    use_nodes: bool = True
+    linear_execution: bool = False
 
 
 class SimulationBuilder:
@@ -387,45 +58,21 @@ class SimulationBuilder:
         parameter_grid: ParameterGrid,
         out_file: str,
         import_line: str,
-        n_cores: int = 1,
-        n_cores_to_use: int = 1,
-        n_gather_cores: int = 1,
-        n_gather_per_core: int = 1,
-        use_nodes: bool = False,
-        runs_per_node: Optional[int] = None,
-        linear_execution: bool = False,
-        hrs: Optional[int] = None,
-        min: Optional[int] = None,
-        sec: Optional[int] = None,
+        simulation_config: SimulationConfig,
+        gather_config: Optional[GatherConfig] = None,
     ):
 
         self._has_complete_params: bool = False
 
         self._import_line: str = import_line
 
-        self._n_cores: int = n_cores
+        self._simulation_config: SimulationConfig = simulation_config
 
-        self._n_cores_to_use: int = n_cores_to_use
-
-        self._n_gather_per_core: int = n_gather_per_core
-
-        self._n_gather_cores: int = n_gather_cores
-
-        self._use_nodes: bool = use_nodes
-
-        self._runs_per_node: Optional[int] = runs_per_node
+        self._gather_config: Optional[GatherConfig] = gather_config
 
         self._out_file: str = out_file
 
         self._base_dir: Path = Path(out_file).parent.absolute()
-
-        self._linear_execution: bool = linear_execution
-
-        self._hrs: Optional[int] = hrs
-
-        self._min: Optional[int] = min
-
-        self._sec: Optional[int] = sec
 
         # write out the parameter file
 
@@ -445,7 +92,7 @@ class SimulationBuilder:
         # we need to see how many
         # we need
 
-        if self._use_nodes:
+        if self._simulation_config.use_nodes:
 
             self._compute_chunks()
 
@@ -457,14 +104,47 @@ class SimulationBuilder:
 
         self._generate_python_script()
 
-        if self._use_nodes:
+        if self._simulation_config.use_nodes:
 
             self._generate_slurm_script()
 
     @classmethod
     def from_yaml(cls, file_name: str) -> "SimulationBuilder":
-        pass
 
+        with Path(file_name).open("r") as f:
+
+            inputs = yaml.load(stream=f, Loader=yaml.SafeLoader)
+
+        parameter_grid = ParameterGrid.from_yaml(inputs.pop("parameter_grid"))
+
+        simulation_input = inputs.pop("simulation")
+
+        if "time" in simulation_input:
+
+            sim_time = SLURMTime(**simulation_input.pop("time"))
+
+        else:
+
+            sim_time = SLURMTime()
+
+        simulation_config = SimulationConfig(time=sim_time, **simulation_input)
+
+        gather_inputs = None
+
+        if "gather" in inputs:
+
+            gather_inputs = inputs.pop("gather")
+
+            gather_time = SLURMTime(**gather_inputs.pop("time"))
+
+            gather_config = GatherConfig(time=gather_time, **gather_inputs)
+
+        return cls(
+            parameter_grid=parameter_grid,
+            simulation_config=simulation_config,
+            gather_config=gather_config,
+            **inputs,
+        )
 
     def _initialize_database(self) -> None:
 
@@ -566,23 +246,25 @@ class SimulationBuilder:
 
     def _compute_chunks(self) -> None:
 
-        if self._runs_per_node is None:
+        if self._simulation_config.run_per_node == 1:
 
             runs_per_node = 1
 
-            generator = range(self._n_cores)
+            generator = range(self._simulation_config.n_mp_jobs)
 
-            n_nodes = np.ceil(self._n_iterations / self._n_cores)
+            n_nodes = np.ceil(
+                self._n_iterations / self._simulation_config.n_mp_jobs
+            )
 
         else:
 
-            runs_per_node = self._runs_per_node
+            runs_per_node = self._simulation_config.run_per_node
 
             generator = range(runs_per_node)
 
             n_nodes = np.ceil(self._n_iterations / runs_per_node)
 
-        if self._use_nodes:
+        if self._simulation_config.use_nodes:
 
             self._n_nodes = int(n_nodes)
 
@@ -612,12 +294,15 @@ class SimulationBuilder:
 
         # now collect the gather information
 
-        if self._use_nodes:
+        if self._simulation_config.use_nodes:
 
             self._n_gather_nodes = int(
                 np.ceil(
                     self._n_iterations
-                    / (self._n_gather_cores * self._n_gather_per_core)
+                    / (
+                        self._gather_config.n_cores_per_node
+                        * self._gather_config.n_gather_per_core
+                    )
                 )
             )
 
@@ -626,15 +311,17 @@ class SimulationBuilder:
 
             log.debug(f"number nodes: {self._n_gather_nodes}")
             log.debug(
-                f"total_ranks: {self._n_gather_nodes * self._n_gather_cores}"
+                f"total_ranks: {self._n_gather_nodes * self._gather_config.n_cores_per_node}"
             )
             log.debug(f"number iterations: {self._n_iterations}")
 
-            for i in range(self._n_gather_nodes * self._n_gather_cores):
+            for i in range(
+                self._n_gather_nodes * self._gather_config.n_cores_per_node
+            ):
 
                 core_list = []
 
-                for j in range(self._n_gather_per_core):
+                for j in range(self._gather_config.n_gather_per_core):
 
                     if n < self._n_iterations:
 
@@ -656,9 +343,9 @@ class SimulationBuilder:
             str(self._parameter_file),
             self._base_dir,
             self._import_line,
-            self._n_cores,
+            self._simulation_config.n_mp_jobs,
             self._n_nodes,
-            self._linear_execution,
+            self._simulation_config.linear_execution,
             self._has_complete_params,
             self._current_database_size,
         )
@@ -669,23 +356,23 @@ class SimulationBuilder:
 
         slurm_gen: SLURMGenerator = SLURMGenerator(
             "run_simulation.sh",
-            self._n_cores,
-            self._n_cores_to_use,
+            self._simulation_config.n_mp_jobs,
+            self._simulation_config.n_cores_per_node,
             self._n_nodes,
-            self._hrs,
-            self._min,
-            self._sec,
+            self._simulation_config.time.hrs,
+            self._simulation_config.time.min,
+            self._simulation_config.time.sec,
         )
 
         slurm_gen.write(str(self._base_dir))
 
         slurm_gen: SLURMGatherGenerator = SLURMGatherGenerator(
             "gather_results.sh",
-            self._n_gather_cores,
+            self._gather_config.n_cores_per_node,
             self._n_gather_nodes,
-            self._hrs,
-            self._min,
-            self._sec,
+            self._gather_config.time.hrs,
+            self._gather_config.time.min,
+            self._gather_config.time.sec,
         )
 
         slurm_gen.write(str(self._base_dir))
